requirements.txt
# Core ML
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0

# Whisper fine-tuning
openai-whisper>=20231117
faster-whisper>=0.10.0
evaluate>=0.4.0
jiwer>=3.0.0  # WER calculation

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
webrtcvad>=2.0.10

# API
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Development
pytest>=7.4.0
black>=23.12.0
ruff>=0.1.0
ipykernel>=6.28.0

# Optional: ElevenLabs for synthetic data generation
elevenlabs>=0.2.0
